Add files via upload
Added a GUI for the BB plots, currently it is separate from the GUI for the CCSB calculation. The GUI takes user input and passes it to the BB functions which then add it to the graph so the user can see where abouts on the plot their protein belongs, it will infer as whether the protein is likely disordered or ordered in structure.